Admin dashboard layout

+++ src/app/admin/journalists/journalist-form/journalist-form.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { ActivatedRoute } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 import { User } from 'src/app/security/models/user.model';
 import { UserService } from 'src/app/security/services/user.service';
 
@@ -17,7 +17,7 @@ export class JournalistFormComponent implements OnInit {
 
   journalist: User = new User(0, '', '', '', '', '', 2);
 
-  constructor(private _userService: UserService, private snackBar: MatSnackBar, private route: ActivatedRoute,) {
+  constructor(private _userService: UserService, private snackBar: MatSnackBar, private route: ActivatedRoute, private router: Router) {
     if (this.route.snapshot.paramMap.get('id')) {
       this.update = true;
       const journalistID = this.route.snapshot.paramMap.get('id');
@@ -45,8 +45,9 @@ export class JournalistFormComponent implements OnInit {
           console.log("Add journalist result:", result);
           if (result) {
             this.journalist.userID = result.userID; // Save the real ID
-            console.log("Journamist is added");
-            this.openSnackBar(("Journalist '" + this.journalist.firstName + " " + this.journalist.lastName + "' is saved!"), "Undo")
+            console.log("Journalist is added");
+            this.openSnackBar(("Journalist '" + this.journalist.firstName + " " + this.journalist.lastName + "' is saved!"), "Undo");
+            this.router.navigate(['/journalists']);
           }
 
         }
@@ -60,7 +61,8 @@ export class JournalistFormComponent implements OnInit {
           // Handle result
           console.log("Update journalist result:", result);
           console.log("Journalist is updated");
-          this.openSnackBar(("Journalist '" + this.journalist.firstName + " " + this.journalist.lastName + "' is saved!"), "Undo")
+          this.openSnackBar(("Journalist '" + this.journalist.firstName + " " + this.journalist.lastName + "' is saved!"), "")
+          this.router.navigate(['/journalists']);
         },
         error => {
           console.log("Error:", error);
@@ -82,14 +84,16 @@ export class JournalistFormComponent implements OnInit {
       console.log("The snackbar action was triggerd");
 
       if (action != "Dismiss") {
+        if (!this.update) {
+          // Undo the action --> delete the newly created journalist
+          this._userService.deleteUser(this.journalist.userID).subscribe(
+            () => {
+              console.log("Action is undone, new journalist is deleted");
+              this.snackBar.open("Action is undone", 'Dismiss', { duration: 3000 });
+            }
+          );
+        }
 
-        // Undo the action --> delete the newly created journalist
-        this._userService.deleteUser(this.journalist.userID).subscribe(
-          () => {
-            console.log("Action is undone, new journalist is deleted");
-            this.snackBar.open("Action is undone", 'Dismiss', { duration: 3000 });
-          }
-        );
       }
     });
